Updated to Bevy 0.10.1
Replaced stages with stageless sets. Replaced the function body of update_mouse_position to handle Window entities.

+++ bracket-bevy/src/consoles/update_system.rs
@@ -7,6 +7,8 @@ use bevy::{
     sprite::Mesh2dHandle,
     window::WindowResized,
 };
+use bevy::utils::petgraph::visit::NodeRef;
+use bevy::window::{PrimaryWindow, WindowRef};
 
 use super::{BracketMesh, ScreenScaler};
 
@@ -95,35 +97,56 @@ pub(crate) fn apply_all_batches(mut context: ResMut<BracketContext>) {
 }
 
 pub(crate) fn update_mouse_position(
-    wnds: Res<Windows>,
+    wnds: Query<&Window>,
+    primary_window: Query<&Window, With<PrimaryWindow>>,
     q_camera: Query<(&Camera, &GlobalTransform), With<BracketCamera>>,
     mut context: ResMut<BracketContext>,
-    scaler: Res<ScreenScaler>,
+    mut scaler: ResMut<ScreenScaler>,
 ) {
     // Modified from: https://bevy-cheatbook.github.io/cookbook/cursor2world.html
     // Bevy really needs a nicer way to do this
     let (camera, camera_transform) = q_camera.single();
-    let wnd = if let RenderTarget::Window(id) = camera.target {
-        wnds.get(id)
-    } else {
-        wnds.get_primary()
-    };
+    
+    let mut wnd_opt: Option<&Window> = None;
+    
+    match camera.target {
+        RenderTarget::Window(window_ref) => {
+            match window_ref {
+                WindowRef::Primary => {
+                    wnd_opt =Some(primary_window.single());
+                }
+                
+                WindowRef::Entity(entity) => {
+                    wnd_opt = Some(wnds.get(entity)
+                              .expect(format!("Couldn't get window with entity id {} and generation {} from query!", 
+                                              entity.index(),
+                                              entity.generation()).as_str())
+                    );
+                }
+            }
+        }
+        // Is there a reason why the camera's RenderTarget should ever be an image??
+        RenderTarget::Image(_) => { panic!("The camera RenderTarget was an image, we shouldn't reach this case!") }
+    }
+    
+    if wnd_opt.is_some() {
+        let wnd = wnd_opt.expect("Window was None!");
+        if let Some(screen_pos) = wnd.cursor_position() {
+            
+            // This is a workaround to ensure that the Scaler's width and height are never 0.
+            // Res<Scaler> had to be changed to ResMut<Scaler>
+            // We should probably check why it's occasionally 0 to begin with.
+            scaler.set_screen_size(wnd.width(), wnd.height());
+            
+            let window_size = Vec2::new(wnd.width() as f32, wnd.height() as f32);
+            let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
+            let ndc_to_world = camera_transform.compute_matrix() * camera.projection_matrix().inverse();
+            let world_pos = ndc_to_world.project_point3(ndc.extend(-1.0));
+            let world_pos: Vec2 = world_pos.truncate();
 
-    let wnd = if let Some(wnd) = wnd {
-        wnd
-    } else {
-        return;
-    };
+            let result = (world_pos.x, world_pos.y);
 
-    if let Some(screen_pos) = wnd.cursor_position() {
-        let window_size = Vec2::new(wnd.width() as f32, wnd.height() as f32);
-        let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
-        let ndc_to_world = camera_transform.compute_matrix() * camera.projection_matrix().inverse();
-        let world_pos = ndc_to_world.project_point3(ndc.extend(-1.0));
-        let world_pos: Vec2 = world_pos.truncate();
-
-        let result = (world_pos.x, world_pos.y);
-
-        context.set_mouse_pixel_position(result, &scaler);
+            context.set_mouse_pixel_position(result, &scaler);
+        }
     }
 }
